MDL-12730 new definition of debug levels, E_STRICT for DEBUG_DEVELOPER
Originally we used numbers because PHP4 did not understand new constants. We use the constants now because future versions of PHP may change their values. The 32768 was originally intended as a "show debugging" flag, but it was later changed to > comparison in debugging(). This comparison is not technically correct, luckily our debug levels are defined in ascending order so that usually works.