MDL-31096 MDL-31216 Change parameters for URL specification to object
This change simplifies the way in which the inclusion can be called and
additionally allows for inclusion of additional parameters which will be
passed straight through in the ajax post.

This may be particularly useful for contrib plugins such as mod_subpage
which require the subpage ID to be passed through when editing.