MDL-52127 eslint: switch to a more YUI tolerant default config
Instead of having the stricter configuration for all files, use
the less strict default config and switch to the stricter config
for building AMD modules.

This means that the eslint commandline/editor inspections will work
better for people, but since grunt is required to build AMD modules
we still get the stricter checking for those files on build.